Move `naturalOrderComparison` from fileSortFunctions into a textUtils helper. Use `naturalOrderComparison` for comparison of collaborator display names.
